requirements.txt
flask>=2.3.0
requests>=2.31.0
Pillow>=10.0.0
openai>=1.0.0
